move installJoinBlock logic to customBlocks
