+++ apps/src/blockly/customBlocks/cdoBlockly/index.js
@@ -1,7 +1,9 @@
 // This file exports customized blocks for CDO Blockly labs.
 
+import {blocks as commonBlocks} from './commonBlocks';
 import {blocks as spritelabBlocks} from './spritelabBlocks';
 
 export default {
+  ...commonBlocks,
   ...spritelabBlocks,
 };
