+++ apps/src/blockly/customBlocks/googleBlockly/index.js
@@ -1,7 +1,9 @@
 // This file exports customized blocks for Google Blockly labs.
 
+import {blocks as commonBlocks} from './commonBlocks';
 import {blocks as spritelabBlocks} from './spritelabBlocks';
 
 export default {
+  ...commonBlocks,
   ...spritelabBlocks,
 };
